backend/requirements.txt
# Crear entorno virtual (.venv)
# pip install -r requirements.txt
# Luego de definir todas las dependencias correctamente
# ejecutar: pip freeze > requirements.txt
# (esto coloca dependencias con versiones en este archivo)

fastapi[standard]
uvicorn[standard]
sqlalchemy
sqlmodel # Agregado para usar SQLModel
pymysql
# alembic
pydantic-settings
python-dotenv
# reportlab